requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
av==11.0.0
beautifulsoup4==4.12.2
bitarray==2.8.2
bitstring==4.1.2
blinker==1.6.3
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
construct==2.8.8
cryptography==41.0.4
cssselect==1.2.0
cssselect2==0.7.0
decorator==4.4.2
Deprecated==1.2.14
deprecation==2.1.0
distro==1.8.0
docopt==0.6.2
dropbox==11.36.2
ecdsa==0.18.0
esptool==4.6.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
folder-tree-generator==0.1.14
fonttools==4.46.0
forex-python==1.8
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
h11==0.14.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
img2pdf==0.5.1
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
loguru==0.7.2
look-like-scanned==1.0.2
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
ocrmypdf==16.0.3
openai==1.1.1
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pdf2image==1.16.3
pdfkit==1.0.0
pdfminer.six==20231228
pdfrw==0.4
piexif==1.1.3
pika==1.3.2
pikepdf==8.11.2
Pillow==10.0.1
playwright==1.40.0
pluggy==1.3.0
ply==3.11
proglog==0.1.10
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pydyf==0.8.0
pyee==11.0.1
pyexiv2==2.5.0
Pygments==2.16.1
pymp4==1.4.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.25.0
pyphen==0.14.0
pypng==0.20220715.0
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytube==15.0.0
pytz==2023.3.post1
PyVirtualDisplay==3.0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
reedsolo==1.7.0
regex==2023.10.3
reportlab==4.0.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
selenium==4.15.2
sgmllib3k==1.0.0
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stone==3.3.1
text-unidecode==1.3
tiktoken==0.5.1
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.0
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.0.6
weasyprint==60.2
webencodings==0.5.1
Werkzeug==3.0.1
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
youtube-transcript-api==0.6.1
zopfli==0.2.3
